requirements.txt
alembic==1.9.1
aniso8601==9.0.1
bidict==0.22.0
click==8.1.3
dnspython==1.16.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-CKEditor==0.4.6
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.2
flask-talisman==1.0.0
Flask-WTF==1.0.1
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
WTForms==3.0.1
zope.event==4.6
zope.interface==5.5.2
